refactor(dockets): Move functions to fix circular import
