+++ cl/opinion_page/views.py
@@ -1,14 +1,13 @@
 import datetime
 from collections import OrderedDict, defaultdict
 from itertools import groupby
-from typing import Dict, Tuple, Union
+from typing import Dict, Union
 from urllib.parse import urlencode
 
 import eyecite
 import natsort
 from asgiref.sync import sync_to_async
 from django.contrib import messages
-from django.contrib.auth.models import AnonymousUser, User
 from django.core.exceptions import ObjectDoesNotExist, PermissionDenied
 from django.core.paginator import EmptyPage, PageNotAnInteger, Paginator
 from django.db.models import F, IntegerField, Prefetch
@@ -31,7 +30,6 @@ from reporters_db import (
 )
 from rest_framework.status import HTTP_300_MULTIPLE_CHOICES, HTTP_404_NOT_FOUND
 
-from cl.alerts.models import DocketAlert
 from cl.citations.parenthetical_utils import get_or_create_parenthetical_groups
 from cl.custom_filters.templatetags.text_filters import best_case_name
 from cl.favorites.forms import NoteForm
@@ -55,6 +53,7 @@ from cl.opinion_page.forms import (
     CourtUploadForm,
     DocketEntryFilterForm,
 )
+from cl.opinion_page.utils import core_docket_data, get_case_title
 from cl.people_db.models import AttorneyOrganization, CriminalCount, Role
 from cl.recap.constants import COURT_TIMEZONES
 from cl.search.models import (
@@ -192,69 +191,6 @@ async def redirect_docket_recap(
     )
     return HttpResponseRedirect(
         reverse("view_docket", args=[docket.pk, docket.slug])
-    )
-
-
-def get_case_title(cluster: OpinionCluster) -> str:
-    return f"{trunc(best_case_name(cluster), 100)}, {cluster.citation_string}"
-
-
-def make_docket_title(docket: Docket) -> str:
-    title = ", ".join(
-        [
-            s
-            for s in [
-                trunc(best_case_name(docket), 100, ellipsis="..."),
-                docket.docket_number,
-            ]
-            if s and s.strip()
-        ]
-    )
-    return title
-
-
-def user_has_alert(user: Union[AnonymousUser, User], docket: Docket) -> bool:
-    has_alert = False
-    if user.is_authenticated:
-        has_alert = DocketAlert.objects.filter(
-            docket=docket, user=user, alert_type=DocketAlert.SUBSCRIPTION
-        ).exists()
-    return has_alert
-
-
-def core_docket_data(
-    request: HttpRequest,
-    pk: int,
-) -> Tuple[Docket, Dict[str, Union[bool, str, Docket, NoteForm]]]:
-    """Gather the core data for a docket, party, or IDB page."""
-    docket = get_object_or_404(Docket, pk=pk)
-    title = make_docket_title(docket)
-
-    try:
-        note = Note.objects.get(docket_id=docket.pk, user=request.user)
-    except (ObjectDoesNotExist, TypeError):
-        # Not saved in notes or anonymous user
-        note_form = NoteForm(
-            initial={
-                "docket_id": docket.pk,
-                "name": trunc(best_case_name(docket), 100, ellipsis="..."),
-            }
-        )
-    else:
-        note_form = NoteForm(instance=note)
-
-    has_alert = user_has_alert(request.user, docket)
-
-    return (
-        docket,
-        {
-            "docket": docket,
-            "title": title,
-            "note_form": note_form,
-            "has_alert": has_alert,
-            "timezone": COURT_TIMEZONES.get(docket.court_id, "US/Eastern"),
-            "private": docket.blocked,
-        },
     )
